# hw12_13_14_15_16_calendar/internal/server/http/api/types.gen.go
// Package api provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/deepmap/oapi-codegen version v1.16.3 DO NOT EDIT.
package api

import (
	"time"
)

// CreateEventRequest defines model for CreateEventRequest.
type CreateEventRequest struct {
	// Description Описание события
	Description *string `json:"description,omitempty"`

	// EndTime Время окончания события
	EndTime time.Time `json:"end_time"`

	// NotifyBefore За сколько секунд уведомить о событии
	NotifyBefore *int `json:"notify_before,omitempty"`

	// StartTime Время начала события
	StartTime time.Time `json:"start_time"`

	// Title Заголовок события
	Title string `json:"title"`

	// UserId ID пользователя
	UserId string `json:"user_id"`
}

// ErrorResponse defines model for ErrorResponse.
type ErrorResponse struct {
	Code    *int    `json:"code,omitempty"`
	Error   *string `json:"error,omitempty"`
	Message *string `json:"message,omitempty"`
}

// Event defines model for Event.
type Event struct {
	// Description Описание события
	Description *string `json:"description,omitempty"`

	// EndTime Время окончания события
	EndTime time.Time `json:"end_time"`

	// Id Уникальный идентификатор события
	Id string `json:"id"`

	// NotifyBefore За сколько секунд уведомить о событии
	NotifyBefore *int `json:"notify_before,omitempty"`

	// StartTime Время начала события
	StartTime time.Time `json:"start_time"`

	// Title Заголовок события
	Title string `json:"title"`

	// UserId ID пользователя
	UserId string `json:"user_id"`
}

// SuccessResponse defines model for SuccessResponse.
type SuccessResponse struct {
	Message *string `json:"message,omitempty"`
	Success *bool   `json:"success,omitempty"`
}

// UpdateEventRequest defines model for UpdateEventRequest.
type UpdateEventRequest struct {
	// Description Описание события
	Description *string `json:"description,omitempty"`

	// EndTime Время окончания события
	EndTime time.Time `json:"end_time"`

	// NotifyBefore За сколько секунд уведомить о событии
	NotifyBefore *int `json:"notify_before,omitempty"`

	// StartTime Время начала события
	StartTime time.Time `json:"start_time"`

	// Title Заголовок события
	Title string `json:"title"`

	// UserId ID пользователя
	UserId string `json:"user_id"`
}

// BadRequest defines model for BadRequest.
type BadRequest = ErrorResponse

// InternalError defines model for InternalError.
type InternalError = ErrorResponse

// NotFound defines model for NotFound.
type NotFound = ErrorResponse

// CreateEventJSONRequestBody defines body for CreateEvent for application/json ContentType.
type CreateEventJSONRequestBody = CreateEventRequest

// UpdateEventJSONRequestBody defines body for UpdateEvent for application/json ContentType.
type UpdateEventJSONRequestBody = UpdateEventRequest